# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SlotIQ Scheduling API",
    "version": "1.0.0",
    "description": "Natural-language meeting scheduling, deterministic availability analysis, agenda generation, and meeting history."
  },
  "servers": [{ "url": "/" }],
  "paths": {
    "/api/parse-meeting-request": {
      "post": {
        "summary": "Parse a meeting request",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ParseRequest" } } }
        },
        "responses": { "200": { "description": "Parsed constraints or a clarification question" }, "400": { "$ref": "#/components/responses/BadRequest" } }
      }
    },
...
  },
  "components": {
    "responses": { "BadRequest": { "description": "Invalid request" } },
    "schemas": {
      "ParseRequest": {
        "type": "object",
        "required": ["input"],
        "properties": {
          "input": { "type": "string" },
          "timezone": { "type": "string" },
          "clarificationState": { "type": "object", "properties": { "baseInput": { "type": "string" }, "step": { "type": "string", "enum": ["duration", "attendees"] } } }
        }
      },
...
    }
  }
}